+++ apps/web/src/features/tracks/components/track-card.tsx
@@ -1,7 +1,6 @@
 import { memo } from 'react';
 
-import { Card, CardContent, CardDescription, CardTitle } from '@/components/ui/card';
-import { Badge } from '@/components/ui/badge';
+import { Card, CardContent } from '@/components/ui/card';
 import { Track } from '@/types/entities/track';
 
 import TrackAudio from '@/features/tracks/components/track-audio';
@@ -10,22 +9,13 @@ import TrackUpload from '@/features/tracks/components/track-upload';
 import TrackActions from '@/features/tracks/components/track-actions';
 import TrackSelector from '@/features/tracks/components/track-selector';
 
-import useDeleteTrackMutation from '@/features/tracks/hooks/use-delete-track-mutation';
-import useTracksStore from '@/features/tracks/store/use-tracks.store';
-import useTrackStore from '@/features/tracks/store/use-track.store';
+import TrackInfo from '@/features/tracks/components/track-info';
 
 interface TrackCardProps {
   track: Track;
 }
 
 const TrackCard = memo(({ track }: TrackCardProps) => {
-  const activeTrack = useTrackStore((state) => state.activeTrack);
-
-  const trackTitle =
-    activeTrack?.id === track.id
-      ? (activeTrack.title ?? track.title)
-      : track.title;
-
   return (
     <Card
       data-testid={`track-item-${track.id}`}
@@ -40,41 +30,7 @@ const TrackCard = memo(({ track }: TrackCardProps) => {
       />
 
       <CardContent className="pb-6 h-full flex flex-col justify-between">
-        <div className="mb-4">
-          <CardTitle
-            data-testid={`track-item-${track.id}-title`}
-            className="text-xl truncate text-center"
-          >
-            {trackTitle}
-          </CardTitle>
-          <CardDescription className="text-center text-muted-foreground">
-            <p
-              data-testid={`track-item-${track.id}-album`}
-              className="truncate"
-            >
-              {track.album}
-            </p>
-            <p
-              data-testid={`track-item-${track.id}-artist`}
-              className="truncate"
-            >
-              By {track.artist}
-            </p>
-          </CardDescription>
-
-          <div className="flex flex-wrap items-center justify-center gap-1 mt-2">
-            {track.genres.map((genre, index) => (
-              <Badge
-                data-testid={`genre-${genre}`}
-                key={index}
-                variant="secondary"
-                className="text-xs"
-              >
-                {genre}
-              </Badge>
-            ))}
-          </div>
-        </div>
+        <TrackInfo track={track} />
 
         {track?.audioFile ? (
           <TrackAudio track={track} />
