+++ apps/web/src/features/tracks/components/track-info.tsx
@@ -2,19 +2,27 @@ import { Badge } from '@/components/ui/badge';
 import { CardDescription, CardTitle } from '@/components/ui/card';
 import { Track } from '@/types/entities/track';
 import { memo } from 'react';
+import useTrackStore from '@/features/tracks/store/use-track.store';
 
 interface TrackInfoProps {
   track: Track;
 }
 
 const TrackInfo = memo(({ track }: TrackInfoProps) => {
+  const activeTrack = useTrackStore((state) => state.activeTrack);
+
+  const trackTitle =
+    activeTrack?.id === track.id
+      ? (activeTrack.title ?? track.title)
+      : track.title;
+
   return (
     <div className="mb-4">
       <CardTitle
         data-testid={`track-item-${track.id}-title`}
         className="text-xl truncate text-center"
       >
-        {track.title}
+        {trackTitle}
       </CardTitle>
       <CardDescription className="text-center text-muted-foreground">
         <p data-testid={`track-item-${track.id}-album`} className="truncate">
